Implemented sweet alert and looged admin user account not be deleted or edited

+++ app/Http/Controllers/Admin/RoleUserController.php
@@ -32,10 +32,25 @@ class RoleUserController extends BaseController
      */
     public function edit($id)
     {
+        
+        // finding the current admin user.
+         $admin = Admin::where('id', $id)->first();             
+         // getting the logged user
+         $logged_user = auth()->user();       
+         //Checking logged admin user 
+         if($logged_user <> $admin ){
         // Attaching pagetitle and subtitle to view.
-        view()->share(['pageTitle' => 'Funville Users', 'subTitle' => 'List of Funville users and roles' ]);        
-        $admin = Admin::where('id', $id)->first();
-        return view('admin.user.role', compact('admin'));
+	        view()->share(['pageTitle' => 'Funville Users', 'subTitle' => 'List of Funville users and roles' ]);        
+        	$admin = Admin::where('id', $id)->first();
+	        return view('admin.user.role', compact('admin'));
+	  }
+	  else{
+
+            // setting flash message using trait
+            $this->setFlashMessage(' Logged Admin User account can\'t be edited', 'info');    
+            $this->showFlashMessages();
+            return redirect()->route('admin.users.index');
+        }
     }
 
     /**
@@ -80,21 +95,33 @@ class RoleUserController extends BaseController
     {
         //finding the user
         $admin = Admin::where('id', $id)->first();
-        // Detach all roles from the user...
-        $admin->roles()->detach();
-        //delete the user.       
-        if($admin->delete()){
-             // setting flash message using trait
-             $this->setFlashMessage(' User account is deleted successfully', 'success');    
-             $this->showFlashMessages();
-              // getting all admin users
-              $admins = Admin::all();
-             // Attaching pagetitle and subtitle to view.
-             view()->share(['pageTitle' => 'Funville Users', 'subTitle' => 'List of Funville users and roles' ]); 
-             return view('admin.user.index', compact('admins'));
+        // getting the logged user
+        $logged_user = auth()->user();       
+        //Checking logged admin user or not
+        if($logged_user <> $admin ){
+        	// Detach all roles from the user...
+        	$admin->roles()->detach();
+        	//delete the user.       
+        	if($admin->delete()){
+             	// setting flash message using trait
+             	$this->setFlashMessage(' User account is deleted successfully', 'success');    
+             	$this->showFlashMessages();
+		return redirect()->route('admin.users.index');
+              	// getting all admin users
+              	//$admins = Admin::all();
+             	// Attaching pagetitle and subtitle to view.
+             	//view()->share(['pageTitle' => 'Funville Users', 'subTitle' => 'List of Funville users and roles' ]); 
+             	//return view('admin.user.index', compact('admins'));
 
-        }else{
-            return $this->responseRedirectBack(' There was an error while deleting the user account' ,'error', false, false); 
+        	}else{
+            	return $this->responseRedirectBack(' There was an error while deleting the user account' ,'error', false, false); 
+        	}
+	}else{
+
+            // setting flash message using trait
+            $this->setFlashMessage(' Logged Admin User account can\'t be deleted', 'info');    
+            $this->showFlashMessages();
+            return redirect()->route('admin.users.index');
         }
 
     }
